% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_affinity_probabilities.R
\name{calculate_affinity_probabilities}
\alias{calculate_affinity_probabilities}
\title{Calculate affinity probability matrix}
\usage{
calculate_affinity_probabilities(total_sets, current_sets, similarity_matrix)
}
\arguments{
\item{total_sets}{Character or numeric vector of all cell identifiers in the network.}

\item{current_sets}{Character or numeric vector of cell identifiers in the current metacell.}

\item{similarity_matrix}{Square numeric matrix representing pairwise similaritys between cells.
Row and column names should match node identifiers in \code{total_sets}.}
}
\value{
A numeric vector of transition probabilities where:
\itemize{
\item Nodes in the \code{current_sets} have affinity 0
\item Other nodes have affinity probabilites proportional to their average similarity to current nodes
}
}
\description{
Computes the affinity probabilities from current metacell (group of cells) to other nodes based on
their average similarity.
}
\examples{
# Create example data
all_cells <- paste0("Cell", 1:100)
current_metacell <- c("Cell1", "Cell5", "Cell10")
sim_mat <- matrix(runif(10000), nrow = 100, dimnames = list(all_cells, all_cells))

# Calculate affinity probabilities
aff_probs <- calculate_affinity_probabilities(
  total_sets = all_cells,
  current_sets = current_metacell,
  similarity_matrix = sim_mat
)
}
