% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_cells_index.R
\name{meta_cells_index}
\alias{meta_cells_index}
\title{Identifiy Cell Indices for Meta-Cell Construction}
\usage{
meta_cells_index(seq_dep, similarity_matrix, meta_cells_num, seed = 123)
}
\arguments{
\item{seq_dep}{Numeric vector of sequencing depths (library sizes) for each cell.}

\item{similarity_matrix}{Square similarity probability between cells.}

\item{meta_cells_num}{Integer specifying the target number meta-cells to create.}

\item{seed}{Integer for random seed. Set for reproducible sampling when creating meta-cells.}
}
\value{
A list where each element contains the indices of cells belonging to one meta-cell.
}
\description{
This function groups individual cells into meta-cells based on sequencing depth and
cell-cell similarity matrix, aiming to create meta-cells with balanced library sizes.
}
\examples{
#Example usage:
seq_depth <- c(1000,1501,800,1200,2000,900,100,700)
similarity <- matrix(runif(64), nrow=8)
meta_indices <- meta_cells_index(seq_depth, similarity, 3)
}
